rename Configuration namespace to ConfigurationSources
